Refactor devserver-specific launch code out of cli
