+++ src/cli/local-commands/run.command.ts
@@ -1,21 +1,12 @@
-import * as webpack from 'webpack'
-import * as devserver from 'webpack-dev-middleware'
-import * as hot from 'webpack-hot-middleware'
-import hotServer  = require('webpack-hot-server-middleware')
-import * as express from 'express'
-import * as path from 'path'
 import * as glob from 'glob'
 import * as dotenv from 'dotenv'
-import errorOverlay = require('react-dev-utils/errorOverlayMiddleware')
-import { pick } from 'lodash'
-import { getWebpackConfig } from '../../lib/bundler/getWebpackConfig'
-import { renderHtmlWrapper } from '../../lib/server/renderHtmlWrapper'
 import { runPluginLoaderHook } from '../../lib/bundler/PluginLoader'
 import { getConfig } from '../../lib/server/getConfig'
 import getImplicitProjectPluginConfigurationsFromFilepaths from '../../lib/server/getProjectPluginConfigs'
+import { startDevserver } from '../../lib/server/startDevServer'
 
 export interface RunCommandOpts {
-  entry: string
+  devServer: boolean
   port: number
 }
 
@@ -27,22 +18,18 @@ export const builder = {
     description: 'Port to start server on',
     default: Number(process.env.PORT || 8000),
   },
-}
-
-export const enforceCustomerFacingHttps: express.RequestHandler = (req, res, next) => {
-  if (req.headers['x-forwarded-proto'] === 'http') {
-    return res.redirect(301, ['https://', req.get('Host'), req.url].join(''))
+  devServer: {
+    alias: 'd',
+    type: 'boolean',
+    description: 'Start in development mode',
+    default: process.env.NODE_ENV !== 'production'
   }
-  return next()
 }
 
-export async function handler({ port }: RunCommandOpts) {
-  const appConfig = getConfig()
-
-  const isProduction = process.env.NODE_ENV === 'production'
-
+export async function handler({ port, devServer }: RunCommandOpts) {
   loadEnvironment()
 
+  const appConfig = getConfig()
   const plugins = {
     ...appConfig.plugins,
     ...getImplicitProjectPluginConfigurationsFromFilepaths(glob.sync(appConfig.projectPlugins))
@@ -50,46 +37,13 @@ export async function handler({ port }: RunCommandOpts) {
 
   await runPluginLoaderHook(plugins, (loader) => loader.applicationWillStart())
 
-  const webpackConfig = getWebpackConfig({
-    pages: getEntrypointFiles(),
-    plugins
+  startDevserver({
+    config: appConfig,
+    plugins,
+    port
   })
-
-  const bundler = webpack(webpackConfig)
-  const app = express()
-
-  const clientBundler = (bundler as any).compilers.find((compiler: webpack.Compiler) => compiler.name === 'client')
-
-  if (isProduction) {
-    app.use(enforceCustomerFacingHttps)
-  }
-
-  app.use(hot(clientBundler, { path: '/_hot' }))
-  app.get('*', devserver(bundler, { noInfo: true, publicPath: '/' }))
-  app.use(hotServer(bundler))
-  app.use(errorOverlay())
-
-  app.get('*', (req, res) => {
-    res.writeHead(200, { 'Content-Type': 'text/html' })
-    res.write(renderHtmlWrapper({ config: getFrontendEnvironment(), devserver: true }))
-    res.end()
-  })
-
-  const server = app.listen(port, () => {
-    const address = `http://localhost:${server.address().port}`
-    process.stderr.write(`Development server started on ${address}\n`)
-  })
-
-  function getEntrypointFiles() {
-    const filePattern = 'src/pages/**/*.@(t|j)s?(x)'
-    return glob.sync(filePattern).map((subpath) => path.resolve(subpath))
-  }
 
   function loadEnvironment() {
     dotenv.config()
-  }
-
-  function getFrontendEnvironment(): NodeJS.ProcessEnv {
-    return pick(process.env, appConfig.frontendEnvironment)
   }
 }
